Updated display_message to do formatting at the same time.
Also updated how to define a color (for example it used to be {{COLOR_RED}}, now it is just {{RED}})

Also made the formatting automatically remove any color from the string at the end (so as to not bleed the color into the next print statement).

Added a finish_refreshable_message function to flush the buffer when the refreshed message is no longer needed.